+++ src/lib/services/evm/index.ts
@@ -4,15 +4,17 @@ import {
   CELATONE_QUERY_KEYS,
   useCelatoneApp,
   useCurrentChain,
+  useEvmConfig,
 } from "lib/app-provider";
-import type { HexAddr20 } from "lib/types";
-import { isHexWalletAddress } from "lib/utils";
+import { type HexAddr20, zHexAddr20 } from "lib/types";
+import { bech32AddressToHex, isHexWalletAddress } from "lib/utils";
 
 import {
   getEvmCodesByAddress,
   getEvmContractInfoSequencer,
   getEvmParams,
 } from "./lcd";
+import { getEthCall } from "./jsonRpc";
 
 export const useEvmParams = () => {
   const {
@@ -65,6 +67,34 @@ export const useEvmContractInfoSequencer = (address: HexAddr20) => {
       retry: 1,
       refetchOnWindowFocus: false,
       enabled: address && isHexWalletAddress(address),
+    }
+  );
+};
+
+export const useEthCall = (to: HexAddr20, data: string) => {
+  const { address } = useCurrentChain();
+  const hexAddr = address
+    ? zHexAddr20.parse(bech32AddressToHex(address))
+    : undefined;
+  const evm = useEvmConfig({ shouldRedirect: false });
+
+  return useQuery(
+    [
+      CELATONE_QUERY_KEYS.EVM_ETH_CALL,
+      evm.enabled && evm.jsonRpc,
+      hexAddr,
+      to,
+      data,
+    ],
+    async () => {
+      if (!evm.enabled) throw new Error("EVM is not enabled (useEthCall)");
+      return getEthCall(evm.jsonRpc, hexAddr ?? null, to, data);
+    },
+    {
+      enabled: evm.enabled && !!evm.jsonRpc,
+      retry: false,
+      refetchOnWindowFocus: false,
+      refetchOnMount: false,
     }
   );
 };